feat: Refactoring architecture
