+++ src/main/java/com/iexapis/cloud/integration/concurrencystock/adapter/rest/configuration/ApiClientConfiguration.java
@@ -1,4 +1,4 @@
-package com.iexapis.cloud.integration.concurrencystock.configuration;
+package com.iexapis.cloud.integration.concurrencystock.adapter.rest.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -12,7 +12,7 @@ import org.springframework.web.client.RestTemplate;
  * 22.09.2023
  */
 @Configuration
-public class WebConfiguration {
+public class ApiClientConfiguration {
 
     @Bean
     public RestTemplate restTemplate(){
