+++ src/main/java/com/iexapis/cloud/integration/concurrencystock/adapter/rest/model/Stock.java
@@ -1,4 +1,4 @@
-package com.iexapis.cloud.integration.concurrencystock.model;
+package com.iexapis.cloud.integration.concurrencystock.adapter.rest.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -9,7 +9,7 @@ import lombok.RequiredArgsConstructor;
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 @RequiredArgsConstructor
-public class StockModel {
+public class Stock {
 
     private String symbol;
 }
