--- src/main/java/com/iexapis/cloud/integration/concurrencystock/api/stock/StockClientApi.java
@@ -1,10 +1,0 @@
-package com.iexapis.cloud.integration.concurrencystock.api.stock;
-
-import com.iexapis.cloud.integration.concurrencystock.model.StockModel;
-
-import java.util.Collection;
-
-public interface StockClientApi {
-
-    Collection<StockModel> getStocks();
-}
